+++ examples/next-openai/app/page.tsx
@@ -4,7 +4,7 @@ import { useChat } from '@ai-sdk/react';
 import ChatInput from '@/component/chat-input';
 
 export default function Chat() {
-  const { error, status, sendMessage, messages, reload, stop } = useChat();
+  const { error, status, sendMessage, messages, regenerate, stop } = useChat();
 
   return (
     <div className="flex flex-col w-full max-w-md py-24 mx-auto stretch">
@@ -38,7 +38,7 @@ export default function Chat() {
           <button
             type="button"
             className="px-4 py-2 mt-4 text-blue-500 border border-blue-500 rounded-md"
-            onClick={() => reload()}
+            onClick={() => regenerate()}
           >
             Retry
           </button>