+++ examples/next-openai/app/use-chat-data-ui-parts/page.tsx
@@ -16,7 +16,7 @@ type MyMessage = UIMessage<
 >;
 
 export default function Chat() {
-  const { error, status, sendMessage, messages, reload, stop } =
+  const { error, status, sendMessage, messages, regenerate, stop } =
     useChat<MyMessage>({
       transport: new DefaultChatTransport({
         api: '/api/use-chat-data-ui-parts',
@@ -84,7 +84,7 @@ export default function Chat() {
           <button
             type="button"
             className="px-4 py-2 mt-4 text-blue-500 border border-blue-500 rounded-md"
-            onClick={() => reload()}
+            onClick={() => regenerate()}
           >
             Retry
           </button>