+++ packages/ai/src/ui/chat.ts
@@ -229,37 +229,31 @@ export abstract class AbstractChat<UI_MESSAGE extends UIMessage> {
     this.state.messages = messages;
   }
 
-  removeAssistantResponse = () => {
-    const lastMessage = this.state.messages[this.state.messages.length - 1];
-
-    if (lastMessage == null) {
-      throw new Error('Cannot remove assistant response from empty chat');
-    }
-
-    if (lastMessage.role !== 'assistant') {
-      throw new Error('Last message is not an assistant message');
-    }
-
-    this.state.popMessage();
-  };
-
   /**
-   * Append a user message to the chat list. This triggers the API call to fetch
+   * Appends or replaces a user message to the chat list. This triggers the API call to fetch
    * the assistant's response.
+   *
+   * If a messageId is provided, the message will be replaced.
    */
   sendMessage = async (
     message:
-      | (CreateUIMessage<UI_MESSAGE> & { text?: never; files?: never })
+      | (CreateUIMessage<UI_MESSAGE> & {
+          text?: never;
+          files?: never;
+          messageId?: string;
+        })
       | {
           text: string;
           files?: FileList | FileUIPart[];
           metadata?: InferUIMessageMetadata<UI_MESSAGE>;
           parts?: never;
+          messageId?: string;
         }
       | {
           files: FileList | FileUIPart[];
           metadata?: InferUIMessageMetadata<UI_MESSAGE>;
           parts?: never;
+          messageId?: string;
         },
     options: ChatRequestOptions = {},
   ): Promise<void> => {
@@ -282,38 +276,85 @@ export abstract class AbstractChat<UI_MESSAGE extends UIMessage> {
       uiMessage = message;
     }
 
-    this.state.pushMessage({
-      ...uiMessage,
-      id: uiMessage.id ?? this.generateId(),
-      role: uiMessage.role ?? 'user',
-    } as UI_MESSAGE);
+    if (message.messageId != null) {
+      const messageIndex = this.state.messages.findIndex(
+        m => m.id === message.messageId,
+      );
 
-    await this.triggerRequest({ requestType: 'generate', ...options });
+      if (messageIndex === -1) {
+        throw new Error(`message with id ${message.messageId} not found`);
+      }
+
+      if (this.state.messages[messageIndex].role !== 'user') {
+        throw new Error(
+          `message with id ${message.messageId} is not a user message`,
+        );
+      }
+
+      // remove all messages after the message with the given id
+      this.state.messages = this.state.messages.slice(0, messageIndex + 1);
+
+      // update the message with the new content
+      this.state.replaceMessage(messageIndex, {
+        ...uiMessage,
+        id: message.messageId,
+        role: uiMessage.role ?? 'user',
+      } as UI_MESSAGE);
+    } else {
+      this.state.pushMessage({
+        ...uiMessage,
+        id: uiMessage.id ?? this.generateId(),
+        role: uiMessage.role ?? 'user',
+      } as UI_MESSAGE);
+    }
+
+    await this.makeRequest({
+      trigger: 'submit-user-message',
+      messageId: message.messageId,
+      ...options,
+    });
   };
 
   /**
-   * Regenerate the last assistant message.
+   * Regenerate the assistant message with the provided message id.
+   * If no message id is provided, the last assistant message will be regenerated.
    */
-  reload = async (options: ChatRequestOptions = {}): Promise<void> => {
-    // TODO stop any ongoing request
-    if (this.lastMessage === undefined) {
-      return;
+  regenerate = async ({
+    messageId,
+    ...options
+  }: {
+    messageId?: string;
+  } & ChatRequestOptions = {}): Promise<void> => {
+    const messageIndex =
+      messageId == null
+        ? this.state.messages.length - 1
+        : this.state.messages.findIndex(message => message.id === messageId);
+
+    if (messageIndex === -1) {
+      throw new Error(`message ${messageId} not found`);
     }
 
-    if (this.lastMessage.role === 'assistant') {
-      this.state.popMessage();
-    }
+    // set the messages to the message before the assistant message
+    this.state.messages = this.state.messages.slice(
+      0,
+      // if the message is a user message, we need to include it in the request:
+      this.messages[messageIndex].role === 'assistant'
+        ? messageIndex
+        : messageIndex + 1,
+    );
 
-    await this.triggerRequest({ requestType: 'generate', ...options });
+    await this.makeRequest({
+      trigger: 'regenerate-assistant-message',
+      messageId,
+      ...options,
+    });
   };
 
   /**
-   * Resume an ongoing chat generation stream. This does not resume an aborted generation.
+   * Attempt to resume an ongoing streaming response.
    */
-  experimental_resume = async (
-    options: ChatRequestOptions = {},
-  ): Promise<void> => {
-    await this.triggerRequest({ requestType: 'resume', ...options });
+  resumeStream = async (options: ChatRequestOptions = {}): Promise<void> => {
+    await this.makeRequest({ trigger: 'resume-stream', ...options });
   };
 
   addToolResult = async ({
@@ -341,8 +382,8 @@ export abstract class AbstractChat<UI_MESSAGE extends UIMessage> {
       const lastMessage = this.lastMessage;
       if (isAssistantMessageWithCompletedToolCalls(lastMessage)) {
         // we do not await this call to avoid a deadlock in the serial job executor; triggerRequest also uses the job executor internally.
-        this.triggerRequest({
-          requestType: 'generate',
+        this.makeRequest({
+          trigger: 'submit-tool-result',
         });
       }
     });
@@ -359,13 +400,19 @@ export abstract class AbstractChat<UI_MESSAGE extends UIMessage> {
     }
   };
 
-  private async triggerRequest({
-    requestType,
+  private async makeRequest({
+    trigger,
     metadata,
     headers,
     body,
+    messageId,
   }: {
-    requestType: 'generate' | 'resume';
+    trigger:
+      | 'submit-user-message'
+      | 'resume-stream'
+      | 'submit-tool-result'
+      | 'regenerate-assistant-message';
+    messageId?: string;
   } & ChatRequestOptions) {
     this.setStatus({ status: 'submitted', error: undefined });
 
@@ -387,7 +434,7 @@ export abstract class AbstractChat<UI_MESSAGE extends UIMessage> {
 
       let stream: ReadableStream<UIMessageStreamPart>;
 
-      if (requestType === 'resume') {
+      if (trigger === 'resume-stream') {
         const reconnect = await this.transport.reconnectToStream({
           chatId: this.id,
           metadata,
@@ -401,13 +448,15 @@ export abstract class AbstractChat<UI_MESSAGE extends UIMessage> {
 
         stream = reconnect;
       } else {
-        stream = await this.transport.submitMessages({
+        stream = await this.transport.sendMessages({
           chatId: this.id,
           messages: this.state.messages,
           abortSignal: activeResponse.abortController.signal,
           metadata,
           headers,
           body,
+          trigger,
+          messageId,
         });
       }
 
@@ -484,11 +533,12 @@ export abstract class AbstractChat<UI_MESSAGE extends UIMessage> {
         messages: this.state.messages,
       })
     ) {
-      await this.triggerRequest({
-        requestType,
+      await this.makeRequest({
         metadata,
         headers,
         body,
+        // secondary requests are triggered by automatic tool execution
+        trigger: 'submit-tool-result',
       });
     }
   }