+++ packages/ai/src/ui/http-chat-transport.ts
@@ -6,20 +6,28 @@ import { UIMessage } from './ui-messages';
 // use function to allow for mocking in tests:
 const getOriginalFetch = () => fetch;
 
-export type PrepareSubmitMessagesRequest<UI_MESSAGE extends UIMessage> =
-  (options: {
+export type PrepareSendMessagesRequest<UI_MESSAGE extends UIMessage> = (
+  options: {
     id: string;
     messages: UI_MESSAGE[];
     requestMetadata: unknown;
     body: Record<string, any> | undefined;
     credentials: RequestCredentials | undefined;
     headers: HeadersInit | undefined;
-  }) => {
-    body: object;
-    headers?: HeadersInit;
-    credentials?: RequestCredentials;
-    api?: string;
-  };
+    api: string;
+  } & {
+    trigger:
+      | 'submit-user-message'
+      | 'submit-tool-result'
+      | 'regenerate-assistant-message';
+    messageId: string | undefined;
+  },
+) => {
+  body: object;
+  headers?: HeadersInit;
+  credentials?: RequestCredentials;
+  api?: string;
+};
 
 export type PrepareReconnectToStreamRequest = (options: {
   id: string;
@@ -27,6 +35,7 @@ export type PrepareReconnectToStreamRequest = (options: {
   body: Record<string, any> | undefined;
   credentials: RequestCredentials | undefined;
   headers: HeadersInit | undefined;
+  api: string;
 }) => {
   headers?: HeadersInit;
   credentials?: RequestCredentials;
@@ -77,7 +86,7 @@ export type HttpChatTransportInitOptions<UI_MESSAGE extends UIMessage> = {
    * @param messages The current messages in the chat.
    * @param requestBody The request body object passed in the chat request.
    */
-  prepareSubmitMessagesRequest?: PrepareSubmitMessagesRequest<UI_MESSAGE>;
+  prepareSendMessagesRequest?: PrepareSendMessagesRequest<UI_MESSAGE>;
 
   prepareReconnectToStreamRequest?: PrepareReconnectToStreamRequest;
 };
@@ -90,7 +99,7 @@ export abstract class HttpChatTransport<UI_MESSAGE extends UIMessage>
   protected headers?: Record<string, string> | Headers;
   protected body?: object;
   protected fetch: FetchFunction;
-  protected prepareSubmitMessagesRequest?: PrepareSubmitMessagesRequest<UI_MESSAGE>;
+  protected prepareSendMessagesRequest?: PrepareSendMessagesRequest<UI_MESSAGE>;
   protected prepareReconnectToStreamRequest?: PrepareReconnectToStreamRequest;
 
   constructor({
@@ -99,7 +108,7 @@ export abstract class HttpChatTransport<UI_MESSAGE extends UIMessage>
     headers,
     body,
     fetch = getOriginalFetch(),
-    prepareSubmitMessagesRequest,
+    prepareSendMessagesRequest,
     prepareReconnectToStreamRequest,
   }: HttpChatTransportInitOptions<UI_MESSAGE>) {
     this.api = api;
@@ -107,21 +116,24 @@ export abstract class HttpChatTransport<UI_MESSAGE extends UIMessage>
     this.headers = headers;
     this.body = body;
     this.fetch = fetch;
-    this.prepareSubmitMessagesRequest = prepareSubmitMessagesRequest;
+    this.prepareSendMessagesRequest = prepareSendMessagesRequest;
     this.prepareReconnectToStreamRequest = prepareReconnectToStreamRequest;
   }
 
-  async submitMessages({
+  async sendMessages({
     abortSignal,
     ...options
-  }: Parameters<ChatTransport<UI_MESSAGE>['submitMessages']>[0]) {
-    const preparedRequest = this.prepareSubmitMessagesRequest?.({
+  }: Parameters<ChatTransport<UI_MESSAGE>['sendMessages']>[0]) {
+    const preparedRequest = this.prepareSendMessagesRequest?.({
+      api: this.api,
       id: options.chatId,
       messages: options.messages,
       body: { ...this.body, ...options.body },
       headers: { ...this.headers, ...options.headers },
       credentials: this.credentials,
       requestMetadata: options.metadata,
+      trigger: options.trigger,
+      messageId: options.messageId,
     });
 
     const api = preparedRequest?.api ?? this.api;
@@ -137,6 +149,8 @@ export abstract class HttpChatTransport<UI_MESSAGE extends UIMessage>
             ...options.body,
             id: options.chatId,
             messages: options.messages,
+            trigger: options.trigger,
+            messageId: options.messageId,
           };
     const credentials = preparedRequest?.credentials ?? this.credentials;
 
@@ -168,6 +182,7 @@ export abstract class HttpChatTransport<UI_MESSAGE extends UIMessage>
     options: Parameters<ChatTransport<UI_MESSAGE>['reconnectToStream']>[0],
   ): Promise<ReadableStream<UIMessageStreamPart> | null> {
     const preparedRequest = this.prepareReconnectToStreamRequest?.({
+      api: this.api,
       id: options.chatId,
       body: { ...this.body, ...options.body },
       headers: { ...this.headers, ...options.headers },