+++ packages/react/src/use-chat.ts
@@ -28,9 +28,9 @@ export type UseChatHelpers<UI_MESSAGE extends UIMessage> = {
 } & Pick<
   AbstractChat<UI_MESSAGE>,
   | 'sendMessage'
-  | 'reload'
+  | 'regenerate'
   | 'stop'
-  | 'experimental_resume'
+  | 'resumeStream'
   | 'addToolResult'
   | 'status'
   | 'messages'
@@ -98,7 +98,7 @@ export function useChat<UI_MESSAGE extends UIMessage = UIMessage>({
 
   useEffect(() => {
     if (resume) {
-      chatRef.current.experimental_resume();
+      chatRef.current.resumeStream();
     }
   }, [resume, chatRef]);
 
@@ -107,10 +107,10 @@ export function useChat<UI_MESSAGE extends UIMessage = UIMessage>({
     messages,
     setMessages,
     sendMessage: chatRef.current.sendMessage,
-    reload: chatRef.current.reload,
+    regenerate: chatRef.current.regenerate,
     stop: chatRef.current.stop,
     error,
-    experimental_resume: chatRef.current.experimental_resume,
+    resumeStream: chatRef.current.resumeStream,
     status,
     addToolResult: chatRef.current.addToolResult,
   };