+++ packages/langchain/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/langchain
 
+## 1.0.0-alpha.15
+
+### Patch Changes
+
+- Updated dependencies [b1e3abd]
+- Updated dependencies [04d5063]
+- Updated dependencies [b4b4bb2]
+- Updated dependencies [d884051]
+- Updated dependencies [954aa73]
+- Updated dependencies [142576e]
+- Updated dependencies [395c85e]
+- Updated dependencies [60e2c56]
+  - ai@5.0.0-alpha.15
+
 ## 1.0.0-alpha.14
 
 ### Patch Changes