+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,20 @@
 
 ### Patch Changes
 
+- Updated dependencies [b1e3abd]
+- Updated dependencies [04d5063]
+- Updated dependencies [b4b4bb2]
+- Updated dependencies [d884051]
+- Updated dependencies [954aa73]
+- Updated dependencies [142576e]
+- Updated dependencies [395c85e]
+- Updated dependencies [60e2c56]
+  - ai@5.0.0-alpha.15
+
+## 0.0.1-alpha.0
+
+### Patch Changes
+
 - Updated dependencies [63f9e9b]
   - ai@5.0.0-alpha.14