+++ packages/vue/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-sdk/vue
 
+## 2.0.0-alpha.15
+
+### Patch Changes
+
+- Updated dependencies [b1e3abd]
+- Updated dependencies [8ba77a7]
+- Updated dependencies [04d5063]
+- Updated dependencies [b4b4bb2]
+- Updated dependencies [d884051]
+- Updated dependencies [954aa73]
+- Updated dependencies [142576e]
+- Updated dependencies [395c85e]
+- Updated dependencies [60e2c56]
+  - ai@5.0.0-alpha.15
+  - @ai-sdk/provider-utils@3.0.0-alpha.15
+
 ## 2.0.0-alpha.14
 
 ### Patch Changes